split module into 'hyper' and 'grpc'
