+++ utils/src/lib.rs
@@ -193,3 +193,5 @@ pub mod fd_budget;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub mod grpc;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod hyper;
